requirements.txt
environs==9.5.0
python-telegram-bot==13.14
google-cloud-api-keys==0.5.1
google-cloud-dialogflow==2.20.0
google-cloud-api-keys==0.5.1
google-cloud-dialogflow==2.20.0
vk-api==11.9.9